Finish off the hungry dead

+++ dnethack-3.4.3/src/makemon.c
@@ -2001,7 +2001,7 @@ register struct	monst	*mtmp;
 			} else if(ptr == &mons[PM_HUNGRY_DEAD]){
 				struct monst *blbtmp;
 				blbtmp = makemon(&mons[PM_BLOB_OF_PRESERVED_ORGANS], mtmp->mx, mtmp->my, MM_ADJACENTOK|MM_NOCOUNTBIRTH);
-				blbtmp->mextra[0] = mtmp->m_id;
+				blbtmp->mextra[0] = (long)mtmp->m_id;
 			}
 		break;
 	    case S_QUANTMECH:
